Move launcher proto / service to internal/verifier
